% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone_energy.R
\name{cyclone_energy}
\alias{cyclone_energy}
\title{Accumulated Cyclone Energy of a given storm}
\usage{
cyclone_energy(storm, year, dat)
}
\arguments{
\item{storm}{name or id of a storm. Storms prior to 1950 require storm id.}

\item{year}{corresponding year in which storm initially occurred}

\item{dat}{data frame containing information of each storm}
}
\value{
a numeric value indicating the accumulated cyclone energy for the
       given storm and year. The value is computed based on the storm's
       max wind each 6 hours excluding max winds less than 35 and excluding
       points where the storm falls outside the Atlantic Basin
}
\description{
Computes the accumulated cyclone energy of a given storm based on its max
wind, location, and time. The ACE is in 10e-04 kn^2 units.
}
\examples{
ace<- cyclone_energy("IVAN", 2004, hurdat)
ace<- cyclone_energy("AL091953", 1953, hurdat)
ace<- cyclone_energy("florence", 1994, hurdat)
}
