% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_track.R
\name{interp_track}
\alias{interp_track}
\title{Interpolated coordinates of a storm's track}
\usage{
interp_track(storm, year, dat)
}
\arguments{
\item{storm}{name or id of a storm. Storms prior to 1950 require storm id.}

\item{year}{corresponding year in which storm initially occurred.}

\item{dat}{data frame containing information of each storm.}
}
\value{
a data frame with the following components:
\itemize{
  \item \code{name}: name of the storm.
  \item \code{date.time}: date and time throughout storm's duration in thirty
   minute increments.
  \item \code{record.id}: used to identify records that correspond to
   landfalls or to indicate the reason for inclusion of a record not at the
   standard synoptic times.
  \item \code{latitude}: interpolated latitude of storm's center, in
   degrees north.
  \item \code{longitude}: interpolated longitude of storm's center. Negative
   values indicate west longitudes, positive values indicate east longitudes.
  \item \code{id}: identification code of chosen storm.
  \item \code{storm.status}: status of storm, referring to level of
   intensity.
  \item \code{max.wind}: the maximum 1-min average wind (in knots)
   associated with the tropical cyclone at an elevation of 10 m with an
   unobstructed exposure.
}
}
\description{
Return data frame containing interpolated latitude and longitude coordinates
of the storm's center at 30 minute increments. Estimated values for selected
variables from the original data set are also included.
}
\examples{
storm <- interp_track("CLAUDETTE", 2021, hurdat)
storm <- interp_track("AL032021", 2021, hurdat)
}
