% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfall.R
\name{landfall}
\alias{landfall}
\title{Determining if a storm made landfall in continental US}
\usage{
landfall(
  storm,
  year,
  dat,
  method = c("landfall.check", "landfall.data", "landfall.strongest")
)
}
\arguments{
\item{storm}{name or id of a storm. Storms prior to 1950 require a storm id.}

\item{year}{corresponding year in which storm initially occurred.}

\item{dat}{data frame containing information of each storm.}

\item{method}{method of choice regarding information to be returned. Methods
include "landfall.check", "landfall.data", "landfall.strongest".}
}
\value{
method "landfall.check" returns TRUE/FALSE indicating
whether storm made landfall in continental US. Method "landfall.data",
returns data frame regarding storm's landfall status at specific times and
positions. Method "landfall.strongest", returns data frame of storm's
strongest landfall along with position and max.wind
}
\description{
Used to either determine if a storm made landfall, produce a storm's
strongest landfall, or generate a data frame containing information
regarding its track and when and where it was over land or sea.
}
\examples{
landfall_check<- landfall("KATRINA", 2005, hurdat, "landfall.check")
landfall_data<- landfall("irene", 2011, hurdat, "landfall.data")
landfall_strongest<- landfall("AL092004", 2004, hurdat, "landfall.strongest")
}
