% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_size.R
\name{storm_size}
\alias{storm_size}
\title{Plots the positions and size of given storms}
\usage{
storm_size(storm.list, year.list, dat, date.time.list, path = TRUE)
}
\arguments{
\item{storm.list}{vector of one or more storm names or ids. Storms prior to
1950 require storm id.}

\item{year.list}{vector of one or more years in which respective storm
initially occurred.}

\item{dat}{data frame containing information for each storm.}

\item{date.time.list}{a vector of date and time values in the
format: Y-m-d H:M:S UTC.}

\item{path}{logical. If TRUE (default), storm path is plotted.}
}
\value{
a \code{ggplot} object displaying map along with position and size of
storm. A red circular shape represents radii of wind at 64 knots. An orange
circular shape represents radii of wind at 50 knots. A green circular shape
represents radii of wind at 34 knots.
}
\description{
Generates a map displaying the position and size of a storm along with its
track if specified. Storm position is determined by its latitude and
longitude, and storm size is determined by its wind radii at 34 knots, 50
knots, and 64 knots.
}
\examples{
size1<- storm_size("KATRINA", 2005, hurdat, "2005-08-28 06:00:00 UTC")
size2<- storm_size("KATRINA", 2005, hurdat, "2005-08-28 06:00:00",
    path=FALSE)
size3<- storm_size(c("KATRINA", "SANDY"), c(2005, 2012), hurdat,
    c("2005-08-28 06:00:00", "2012-10-25 18:00:00"), path=TRUE)
size4<- storm_size(c("AL122005", "AL182012"), c(2005, 2012), hurdat,
    c("2005-08-25 18:00:00", "2012-10-28 12:00:00"), path=FALSE)
}
