% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_map.R
\name{track_map}
\alias{track_map}
\title{Map of storm tracks for a selection of storms}
\usage{
track_map(storm.list, year.list, dat, date.time.list = NULL)
}
\arguments{
\item{storm.list}{A vector of one or more storm names or ids. Storms prior to
1950 require a storm id.}

\item{year.list}{A vector of one or more years in which the respective storm
initially occurred.}

\item{dat}{A data frame containing information about each storm.}

\item{date.time.list}{An optional vector of date and time values in the
format = Y-m-d H:M:S UTC.}
}
\value{
A ggplot displaying the world map along with color coded storm tracks
or specific locations of each inputted storm.
}
\description{
Generates a map displaying the storm tracks or selective points in time for
each given storm
}
\examples{
track_path<- track_map("KATRINA", 2005, hurdat)
track_point<- track_map("KATRINA", 2005, hurdat, "2005-08-23 18:00:00")
track_paths<-track_map(c("KATRINA", "SANDY"), c(2005, 2012), hurdat,
    date.time.list=c("2005-08-23 18:00:00", "2012-10-21 18:00:00 UTC"))
track_points<-track_map(c("KATRINA", "SANDY"), c(2005, 2012), hurdat,
    date.time.list=c("2005-08-23 18:00:00", "2012-10-21 18:00:00 UTC"))
}
